move tokenizer to an upper directory
