+++ flexeval/__init__.py
@@ -1,3 +1,5 @@
+from flexeval.core.tokenizer import *
+
 from .core.chat_dataset import *
 from .core.eval_setups import ChatResponse, EvalSetup, Generation, MultipleChoice, Perplexity
 from .core.evaluate_chat_response import evaluate_chat_response
@@ -10,7 +12,6 @@ from .core.few_shot_generator import *
 from .core.generation_dataset import *
 from .core.language_model import *
 from .core.metric import *
-from .core.metric.tokenizer import *
 from .core.multiple_choice_dataset import *
 from .core.pairwise_comparison import *
 from .core.prompt_template import *
