+++ tests/core/tokenizer/test_whitespace.py
@@ -1,4 +1,4 @@
-from flexeval.core.metric.tokenizer import WhitespaceTokenizer
+from flexeval import WhitespaceTokenizer
 
 
 def test_whitespace_tokenizer() -> None:
